🐛 context 파일에서Firestore 데이터 불러오기
